do not auto re-generate cmake cache

+++ full-linux-build.sh
@@ -1,12 +1,18 @@
-#!/bin/sh
+#!/bin/bash
 
 set -e
 
 python clang-format.py -r pacman/ tests/
-rm -rf build/
-mkdir -p build
-cd build
-cmake ..
+
+if [[ "$1" = '--full' ]]; then
+    rm -rf build/
+    mkdir -p build
+    cd build
+    cmake ..
+    cd ..
+fi
+
+cd build/
 cmake --build .
 ./pacman
 ctest --output-on-failure
